Drain sockets prior to writing to them
OpenSSL does not report connection reset errors on a socket when writing
to it, but for syslog, we're only ever writing to sockets, which means
we never notice a closed socket (at least not until all write buffers
have filled up!).

This fixes the issue by checking sockets for reads before writing to
them (we don't expect to read anything, though). If the socket is
available for reading (which is the case on EOF), then we just drain the
socket, which should normally immediately throw an EOF (and if it
doesn't, we log an error).

In turn, the EOF exception causes us to teardown our socket, then reopen
one.